Video_Based_Left_Behind_Object_and_Threat_Detection/requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
ultralytics>=8.0.0  # YOLOv8 for object detection

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Object Tracking
filterpy>=1.4.5  # Kalman filter for tracking
scipy>=1.11.0

# Activity Recognition
pytorchvideo>=0.1.5
timm>=0.9.0  # PyTorch Image Models

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Video Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# IoT and Communication
paho-mqtt>=1.6.1  # MQTT for ESP32-CAM communication
requests>=2.31.0
flask>=2.3.0  # Web server for receiving ESP32-CAM streams
flask-cors>=4.0.0

# Notifications
twilio>=8.5.0  # SMS notifications
python-telegram-bot>=20.0  # Telegram notifications
smtplib  # Email notifications (built-in)

# Database
sqlalchemy>=2.0.0
sqlite3  # Built-in

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and Monitoring
loguru>=0.7.0
tensorboard>=2.13.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Date and Time
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

